# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resolvy API Documentation",
        "description": "API documentation for the Resolvy helpdesk application",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/v1/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get list of subscriptions",
                "description": "Returns a paginated list of subscriptions",
                "operationId": "getSubscriptionsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubscriptionResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Subscriptions retrieved successfully."
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Create a new subscription",
                "description": "Creates a new subscription record",
                "operationId": "storeSubscription",
                "requestBody": {
                    "description": "Subscription data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/{id}": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get subscription information",
                "description": "Returns subscription data",
                "operationId": "getSubscriptionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of subscription to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Update an existing subscription",
                "description": "Updates a subscription record",
                "operationId": "updateSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of subscription to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Subscription data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Subscription updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Delete an existing subscription",
                "description": "Deletes a subscription record",
                "operationId": "deleteSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of subscription to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerResource": {
                "title": "Customer Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "customer_code": {
                        "type": "string",
                        "readOnly": true,
                        "example": "CUS-0000000001"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "type": {
                        "type": "string",
                        "example": "individual"
                    },
                    "contact_email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "contact_phone": {
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Pagination Meta",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "Pagination Links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/customers?page=2"
                    }
                },
                "type": "object"
            },
...
            "ServicePackageResource": {
                "title": "Service Package Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Gold Tier Support"
                    },
                    "description": {
                        "type": "string",
                        "example": "24/7 phone and email support"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    }
                },
                "type": "object"
            },
            "StoreSubscriptionRequest": {
                "title": "Store Subscription Request",
                "required": [
                    "customer_id",
                    "package_id",
                    "start_date",
                    "end_date",
                    "status"
                ],
                "properties": {
                    "customer_id": {
                        "description": "ID of the customer",
                        "type": "integer",
                        "example": 1
                    },
                    "package_id": {
                        "description": "ID of the service package",
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "description": "Subscription start date",
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-14"
                    },
                    "end_date": {
                        "description": "Subscription end date",
                        "type": "string",
                        "format": "date",
                        "example": "2026-10-14"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "paused",
                            "cancelled",
                            "expired"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "description": "Additional notes",
                        "type": "string",
                        "example": "Initial setup complete."
                    }
                },
                "type": "object"
            },
            "UpdateSubscriptionRequest": {
                "title": "Update Subscription Request",
                "properties": {
                    "customer_id": {
                        "description": "ID of the customer",
                        "type": "integer",
                        "example": 1
                    },
                    "package_id": {
                        "description": "ID of the service package",
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "description": "Subscription start date",
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-14"
                    },
                    "end_date": {
                        "description": "Subscription end date",
                        "type": "string",
                        "format": "date",
                        "example": "2026-10-14"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "paused",
                            "cancelled",
                            "expired"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "description": "Additional notes",
                        "type": "string",
                        "example": "Customer requested a pause."
                    }
                },
                "type": "object"
            },
            "SubscriptionResource": {
                "title": "Subscription Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-14"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-10-14"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Initial setup complete."
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerResource"
                    },
                    "package": {
                        "$ref": "#/components/schemas/ServicePackageResource"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}